Adding in a deploy script and a test of the deployment
